+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
  * Enhance join handling in `Reflects` and `ResultSetHandlers` for improved data retrieval and merging
  * Add pagination support with joins in SQL provider and mappers
+ * Add unit tests for `CaseFormat` conversion functions
 
 ## 1.0.13